+++ encoding/raw/outbound_test.go
@@ -28,7 +28,7 @@ import (
 	"testing"
 
 	"go.uber.org/yarpc"
-	"go.uber.org/yarpc/internal/channel"
+	"go.uber.org/yarpc/internal/clientconfig"
 	"go.uber.org/yarpc/transport"
 	"go.uber.org/yarpc/transport/transporttest"
 
@@ -80,7 +80,7 @@ func TestCall(t *testing.T) {
 
 	for _, tt := range tests {
 		outbound := transporttest.NewMockUnaryOutbound(mockCtrl)
-		client := New(channel.MultiOutbound(caller, service,
+		client := New(clientconfig.MultiOutbound(caller, service,
 			transport.Outbounds{
 				Unary: outbound,
 			}))
@@ -161,7 +161,7 @@ func TestCallOneway(t *testing.T) {
 
 	for _, tt := range tests {
 		outbound := transporttest.NewMockOnewayOutbound(mockCtrl)
-		client := New(channel.MultiOutbound(caller, service,
+		client := New(clientconfig.MultiOutbound(caller, service,
 			transport.Outbounds{
 				Oneway: outbound,
 			}))
@@ -205,7 +205,7 @@ func TestCallOnewayFailure(t *testing.T) {
 	body := []byte{1, 2, 3}
 
 	outbound := transporttest.NewMockOnewayOutbound(mockCtrl)
-	client := New(channel.MultiOutbound(caller, service,
+	client := New(clientconfig.MultiOutbound(caller, service,
 		transport.Outbounds{
 			Oneway: outbound,
 		}))